+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # Changelog
 
+## [0.6.0] - 2019-08-10
+### Changed
+- Refactor menus module.
+
+### Removed
+- Disclaimer message.
+
 ## [0.5.0] - 2019-07-24
 ### Added
-- Initial version
+- Initial version.
 
+[0.6.0]: https://github.com/isylkin/flappy_bird_clone/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/isylkin/flappy_bird_clone/compare/1576c33...v0.5.0